requirements.txt
Flask
Flask-SQLAlchemy
requests
Flask-Migrate
MySQL-python
